+++ app/routes/panelAdmin/product.js
@@ -1,18 +1,24 @@
-const express = require("express");
+const express = require('express');
 const router = express.Router();
-const models = require("../../models");
-const mongoose = require("mongoose");
+const models = require('../../models');
+const mongoose = require('mongoose');
+const statusCode = require('../../values/statusCodes');
 
-router.get("/", async (req, res) => {
+const escapeRegex = (text) => {
+  console.log({ text });
+  return text.replace(/[-[\]{}()*+?.,\\^$|#\s]/g, '\\$&');
+};
+
+router.get('/', async (req, res) => {
   try {
-    const allProducts = await models.Product.find({});
+    const allProducts = await models.Product.find({}).populate('category');
     res.status(200).json(allProducts);
   } catch (error) {
     res.status(500).json({ CODE: 1021 });
   }
 });
 
-router.get("/:productId", async (req, res) => {
+router.get('/:productId', async (req, res) => {
   const productId = req.params.productId;
   console.log(productId.length);
   //TODO Validate object id
@@ -26,24 +32,40 @@ router.get("/:productId", async (req, res) => {
   }
 });
 
+// search product by name
+router.get('/s/:productName', async (req, res) => {
+  try {
+    const regex = new RegExp(escapeRegex(req.params.productName), 'gi');
+    const foundedProducts = await models.Product.find({ name: regex });
+    res.status(200).json(foundedProducts);
+  } catch (error) {
+    console.log({ error });
+    res.status(500).json({ CODE: statusCode.ER_SMT_WRONG });
+  }
+});
+
 // add a new category              /category POST {BODY}
-router.post("/", async (req, res) => {
-  const { name, image, realPrice, weight, unit } = req.body;
-  if ((!name, !image, !realPrice, !weight, !unit))
-    return res.status(400).json({ CODE: 1021 });
+router.post('/', async (req, res) => {
+  const { name, image, realPrice, weight, unit, newPrice, category } = req.body;
   //TODO : check url is our server
   try {
-    await models.Product({ name, image, realPrice, weight, unit }).save();
-    res.status(201).json({ CODE: 2012 });
+    await models.Product({ name, image, realPrice, weight, unit, newPrice, category }).save();
+    res.status(201).json({ CODE: statusCode.AD_PRODUCT });
   } catch (error) {
-    res.status(500).json({ CODE: 1010 });
+    if (error.message.includes('require')) return res.status(500).json({ CODE: statusCode.ER_PARAMS });
+    res.status(500).json({ CODE: statusCode.ER_SMT_WRONG });
   }
 });
 
 //update category by id            /category/:categoryId  {body}
-router.put("/:productId", async (req, res) => {
+router.put('/:productId', async (req, res) => {
   const productId = req.params.productId;
   try {
+    mongoose.Types.ObjectId.isValid(productId);
+
+    // check id in database
+    const foundedProduct = await models.Product.findById(productId);
+    if (!foundedProduct) return res.status(500).json({ CODE: 1058 });
     const { fieldChange, newValue } = req.body;
     const update = {};
     update[fieldChange] = newValue;
@@ -55,17 +77,17 @@ router.put("/:productId", async (req, res) => {
 });
 
 //delete a product by id          /category/:categoryId
-router.delete("/:productId", async (req, res) => {
+router.delete('/:productId', async (req, res) => {
   const productId = req.params.productId;
   try {
     mongoose.Types.ObjectId.isValid(productId);
+    const foundedProduct = await models.Product.findById(productId);
+    if (!foundedProduct) return res.status(500).json({ CODE: 1058 });
     await models.Product.findByIdAndDelete(productId);
-    res.status(204).json({ CODE: 2525 });
+    res.status(200).json({ CODE: statusCode.DL_PRODUCT });
   } catch (error) {
-    res.status(500).json({ CODE: 2424 });
+    res.status(500).json({ CODE: statusCode.ER_SMT_WRONG });
   }
 });
 
 module.exports = router;
-
-
